Completed tasks for #105
 Resolve unused dependencies. See: https://www.bithound.io/github/latticework/jali/master/dependencies/npm
 Resolve @types/node outdated: See: https://www.bithound.io/github/latticework/jali/master/dependencies/npm
 linting disabled. See: https://github.com/latticework/jali/blob/master/.bithoundrc
 esdoc dependency muted. See: https://www.bithound.io/github/latticework/jali/master/dependencies/npm/esdoc
 Enable license checking. See: https://www.bithound.io/settings/github/latticework/jali
 Fix code analysis issues for argument-verifiers.ts. See: https://www.bithound.io/github/latticework/jali/blob/111d6f4e93053fd2646fe9097ce46cc33479e1a0/packages/@jali-ms/util/src/argument-verifiers.ts
 Fix code analysis issues for iterables.ts. See: https://www.bithound.io/github/latticework/jali/blob/111d6f4e93053fd2646fe9097ce46cc33479e1a0/packages/@jali-ms/util/src/iterables.ts
