+++ DESIGN.md
@@ -43,6 +43,11 @@ Linux distros.
 >     - Visual Studio Code configuration
 > - Jali build process
 
+> > ## Parking lot
+> >
+> > - JS module file names: [@rauschma tweet](https://twitter.com/rauschma/status/776184418096521216)
+
+
 <!-- markdownlint-enable ul-style -->
 
 ## Introduction
