+++ cSpell.json
@@ -65,7 +65,9 @@
         "truthy",
         "vagrantfile",
         "validators",
-        "webpackfile"
+        "webpackfile",
+        "typedef",
+        "asiterable"
     ],
     // flagWords - list of words to be always considered incorrect
     // This is useful for offensive words and common spelling errors.
