+++ packages/@jali-ms/util/src/internal.ts
@@ -1,0 +1,2 @@
+export { default as errorMessage } from './internal_errorMessage';
+
