+++ packages/@jali-ms/util/src/internal_errorMessage.ts
@@ -1,0 +1,3 @@
+export default function errorMessage<T>(value: T, message?: string | ((value: T) => string)) {
+  return (typeof message === 'function') ? message(value) : message;
+}
